experiments/tokenizer-comparison/pyproject.toml
[project]
name = "tokenizer-comparison"
version = "0.1.0"
description = "Quantitative and qualitative comparison of tokenization algorithms on scientific papers"
requires-python = ">=3.12"
dependencies = [
    "transformers>=4.44.0",
    "tokenizers>=0.19.0",
    "sentencepiece>=0.2.0",
    "torch>=2.4.0",
    "pymupdf>=1.24.0",
    "pdfminer-six>=20231228",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pydantic>=2.9.0",
    "pandas>=2.2.0",
    "numpy>=1.26.0",
    "matplotlib>=3.9.0",
    "httpx>=0.27.0",
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
]
